requirements.txt
discord.py>=2.3.0
pandas>=2.0.0
matplotlib>=3.7.0
chardet>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
Pillow>=10.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
